+++ src/no_drop/no_drop_empty.rs
@@ -1,5 +1,7 @@
 use std::mem::ManuallyDrop;
 
+pub const DEFAULT_DROP_PANIC_MSG: &str = "Value was dropped without being unwrapped";
+
 /// A wrapper around a `T` value that always [`panic!`]s if dropped without being
 /// [`Self::unwrap`]ed or [`Self::forget`]ten.
 #[derive(
@@ -72,7 +74,7 @@ impl<T> Drop for NoDropEmpty<T> {
     /// [`panic!`]s.
     #[track_caller]
     fn drop(&mut self) {
-        panic!("Value was dropped without being unwrapped");
+        panic!("{}", DEFAULT_DROP_PANIC_MSG);
     }
 }