+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2025-12-10
+
+### Added
+- Conversion Methods
+  - `into_guard` for `DropGuard` types.
+  - `From` `NoDrop` -> `DropGuard` implementations
+  - `TryFrom` `DropGuard` -> `NoDrop` implementations
+  - `GuardNotArmed` error type for `DropGuardPass` -> `NoDrop` conversions
+- `DEFAULT_DROP_PANIC_MSG` constant for default panic messages
+
 ## [0.2.0] - 2025-12-10
 
 ### Added